o Changes made with Orhan, getting to approval of this PR. Remove test_integrate, put those tests in test_grid - one test per .py file in uxarray folder
