+++ uxarray/helpers.py
@@ -8,7 +8,7 @@ import math
 config.DISABLE_JIT = False
 
 
-def parse_grid_type(dataset, **kw):
+def parse_grid_type(dataset):
     """Checks input and contents to determine grid type. Supports detection of
     UGrid, SCRIP, Exodus and shape file.
 
@@ -19,7 +19,7 @@ def parse_grid_type(dataset, **kw):
 
     Returns
     -------
-    mesh_filetype : str
+    mesh_type : str
         File type of the file, ug, exo, scrip or shp
 
     Raises
@@ -31,18 +31,18 @@ def parse_grid_type(dataset, **kw):
     """
     # exodus with coord or coordx
     if "coord" in dataset:
-        mesh_filetype = "exo"
+        mesh_type = "exo"
     elif "coordx" in dataset:
-        mesh_filetype = "exo"
+        mesh_type = "exo"
     # scrip with grid_center_lon
     elif "grid_center_lon" in dataset:
-        mesh_filetype = "scrip"
+        mesh_type = "scrip"
     # ugrid topology
     elif _is_ugrid(dataset):
-        mesh_filetype = "ugrid"
+        mesh_type = "ugrid"
     else:
         raise RuntimeError(f"Could not recognize dataset format.")
-    return mesh_filetype
+    return mesh_type
 
     # check mesh topology and dimension
     try:
@@ -59,7 +59,7 @@ def parse_grid_type(dataset, **kw):
         mesh_topo_dv = ext_ds.filter_by_attrs(cf_role="mesh_topology").keys()
         if list(mesh_topo_dv)[0] != "" and list(topo_dim_dv)[0] != "" and list(
                 face_conn_dv)[0] != "" and list(node_coords_dv)[0] != "":
-            mesh_filetype = "ugrid"
+            mesh_type = "ugrid"
         else:
             raise ValueError(
                 "cf_role is other than mesh_topology, the input NetCDF file is not UGRID format"
@@ -72,14 +72,14 @@ def parse_grid_type(dataset, **kw):
         # check if this is a shp file
         # we won't use xarray to load that file
         if file_extension == ".shp":
-            mesh_filetype = "shp"
+            mesh_type = "shp"
         else:
             msg = str(e) + ': {}'.format(filepath)
     except ValueError as e:
         # check if this is a shp file
         # we won't use xarray to load that file
         if file_extension == ".shp":
-            mesh_filetype = "shp"
+            mesh_type = "shp"
         else:
             msg = str(e) + ': {}'.format(filepath)
     finally:
@@ -88,7 +88,7 @@ def parse_grid_type(dataset, **kw):
                 filepath)
             raise ValueError(msg)
 
-    return mesh_filetype
+    return mesh_type
 
 
 @njit
